refactor(data-model): the fabric base classes get their own directory
`BaseFabricInstance` and `BaseFabricPrimitive` move from `codec-common/` into
a new `fabric-bases/`, with `test/fabric-bases/` alongside and a
`./fabric-bases` export subpath. Neither class imports anything
codec-related, and outside the concrete value classes their consumers are the
top-level utilities `deep-freeze.ts`, `type-check.ts` and `value-hash.ts`.

The new barrel's header states the principle the directory exists to hold,
since nothing in the code carries it: `interface.ts` is the client surface of
the value hierarchy, these bases are the implementer surface, and importing
the former deliberately does not reach the latter. `codec-common/`'s header
loses the paragraph warning that its two kinds of abstract base must not be
counted together; `BaseCodecEngine` is now the only one there.

Measured with `deno info --json src/codecs.ts` over the identical 54-module
graph, runtime edges only: mutually-dependent area pairs go from 3 to 4, the
new one being `(top level) <-> fabric-bases/`. That coupling is real either
way -- the bases reach `interface.ts`, `value-debug.ts` and `deep-freeze.ts`
while `deep-freeze.ts`, `type-check.ts` and `value-hash.ts` reach back -- and
was previously hidden inside `(top level) <-> codec-common/`, which survives
on eleven unrelated edges.

Corrects the formal spec's module-structure table, which placed both bases in
the concrete-class directories: an arrangement the `UnknownValue` /
`ProblematicValue` cycle forbids. The same rows misplaced those two classes,
which live in `codec-common/`, so they are corrected too.
