+++ packages/data-model/src/fabric-bases/BaseFabricPrimitive.ts
