+++ packages/data-model/src/value-hash.ts
@@ -21,7 +21,7 @@ import { utf8SortedKeysOf } from "@commonfabric/utils/utf8";
 import { isDeepFrozen } from "./deep-freeze.ts";
 import { shallowFabricFromNativeValue } from "./native-conversion.ts";
 import { NATIVE_TAGS, tagFromNativeValue } from "./native-type-tags.ts";
-import { BaseFabricInstance } from "@/codec-common/BaseFabricInstance.ts";
+import { BaseFabricInstance } from "@/fabric-bases/BaseFabricInstance.ts";
 import { codecOf } from "@/codec-common/index.ts";
 import { FabricBytes } from "@/fabric-primitives/FabricBytes.ts";
 import { FabricHash } from "@/fabric-primitives/FabricHash.ts";
