+++ packages/data-model/test/fabric-primitives/schema-type.test.ts
@@ -19,7 +19,7 @@ import {
 } from "@commonfabric/api";
 
 import type { FabricPrimitive } from "@/interface.ts";
-import { BaseFabricPrimitive } from "@/codec-common/BaseFabricPrimitive.ts";
+import { BaseFabricPrimitive } from "@/fabric-bases/BaseFabricPrimitive.ts";
 import {
   codecClasses,
   FabricBytes,
